requirements.txt
PySide6==6.6.1 
opencv-python-headless==4.9.0.80 
numpy==1.26.4 
torch==2.2.2 
onnxruntime==1.17.3 
Pillow==10.3.0 
ultralytics==8.1.45
pyinstaller